(refactor): Separate types into *.types.ts files
