+++ jest.config.js
@@ -8,10 +8,4 @@ module.exports = {
   verbose: true,
   collectCoverage: true,
   collectCoverageFrom: ['./src/**'],
-  coveragePathIgnorePatterns: ['./src/main.ts'],
-  coverageThreshold: {
-    global: {
-      lines: 95
-    }
-  }
 }
