+++ src/api.types.ts
@@ -2,7 +2,7 @@ import type {GitHub} from '@actions/github/lib/utils'
 // eslint-disable-next-line import/no-unresolved
 import {GetResponseTypeFromEndpointMethod} from '@octokit/types'
 
-type Octokit = InstanceType<typeof GitHub>
+export type Octokit = InstanceType<typeof GitHub>
 let octokit: Octokit
 
 export interface WorkflowRun {
