+++ src/main.ts
@@ -1,16 +1,13 @@
 import * as core from '@actions/core'
-import {v4 as uuid} from 'uuid'
-import {getConfig, DispatchMethod, ActionOutputs} from './action'
-import * as api from './api'
 import {backOff} from 'exponential-backoff'
+import {v4 as uuid} from 'uuid'
+import {NumberOfAttempts} from './main.types'
+import {getConfig} from './action'
+import {DispatchMethod, ActionOutputs} from './action.types'
+import * as api from './api'
 import {getDispatchedWorkflowRun} from './utils'
 
 const DISTINCT_ID = uuid()
-
-enum NumberOfAttempts {
-  WorkflowId = 3,
-  WorkflowRuns = 5
-}
 
 async function run(): Promise<void> {
   try {
