+++ src/main.types.ts
@@ -1,0 +1,4 @@
+export enum NumberOfAttempts {
+  WorkflowId = 3,
+  WorkflowRuns = 5
+}
